feat: 分页机制的初步实现 - 初步实现基于分页的内存管理器 - 为部分头文件补上保护宏

+++ mm/main.c
@@ -66,10 +66,10 @@ PUBLIC void task_mm()
 			mm_msg.RETVAL = do_kill();
 			break;
 		case MALLOC:
-			mm_msg.RETVAL = do_mallocfree(1);
+			mm_msg.RETVAL = (int)do_mallocfree(1);
 			break;
 		case FREE:
-			mm_msg.RETVAL = do_mallocfree(0);
+			mm_msg.RETVAL = (int)do_mallocfree(0);
 			break;
 		default:
 			dump_msg("MM::unknown msg", &mm_msg);
